requirements.txt
mysql-connector-python==8.2.0
openai-whisper==20231117
torch==2.0.1
torchaudio==2.0.2
requests==2.31.0
python-dotenv==1.0.0
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
